+++ future/moves/http/server.py
@@ -11,6 +11,10 @@ else:
     try:
         from CGIHTTPServer import _url_collapse_path     # needed for a test
     except ImportError:
-        # python 2.7.3
-        from CGIHTTPServer import (
-            _url_collapse_path_split as _url_collapse_path)
+        try:
+            # Python 2.7.0 to 2.7.3
+            from CGIHTTPServer import (
+                _url_collapse_path_split as _url_collapse_path)
+        except ImportError:
+            # Doesn't exist on Python 2.6.x. Ignore it.
+            pass